rename point to textgen
