+++ script/textgen.py
